Rename: first task files
