+++ app/src/main/java/com/app/music_app/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.ui.platform.LocalContext
-import com.app.music_app.study_manager.task_managers.CompareTaskManager
+import com.app.music_app.tasks.logic.CompareTaskManager
 import com.musiclib.notes.Note
 import com.musiclib.notes.note_metadata.NoteName
 import com.musiclib.notes.range.NoteRange
